requirements.txt
docker
docker-compose
django
djangorestframework
psycopg[binary]
